# src/Tools/Refitter.Generator/OpenApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trend.API",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/Dictionary/GetSearchEngines": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "operationId": "GetSearchEngines",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Dictionary/GetContextTypes": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "operationId": "GetContextTypes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Dictionary/GetActiveFilterOptions": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "operationId": "GetActiveFilterOptions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Dictionary/GetSortFilterOptions": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "operationId": "GetSortFilterOptions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "title": "List<KeyValueElementDto>",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyValueElementDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponseModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "KeyValueElementDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:8080/realms/PortfolioRealm/protocol/openid-connect/auth",
            "scopes": { }
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}